+++ src/domains/projects/TopProjects/ProjectItem.tsx
@@ -1,6 +1,7 @@
 import { FC, ReactNode, useState } from 'react'
 import { useTranslation } from 'react-i18next'
-import * as S from './styles'
+import styles from './project.module.css'
+import { motion } from 'framer-motion'
 
 interface IProject {
   project: {
@@ -19,25 +20,32 @@ export const ProjectItem: FC<IProject> = ({ project }) => {
   const { t } = useTranslation()
 
   return (
-    <S.Wrapper
+    <motion.a
+      className={styles.wrapper}
       onHoverStart={() => setIsBeingHovered(true)}
       onHoverEnd={() => setIsBeingHovered(false)}
       href={project.link}
       target='_blank'
       rel='noopener noreferrer'
     >
-      {isBeingHovered && <S.HoverBackground layoutId='background' layout />}
-      <S.Container>
+      {isBeingHovered && (
+        <motion.div
+          className={styles.hoverBackground}
+          layoutId='background'
+          layout
+        />
+      )}
+      <div className={styles.container}>
         <span>
           {project.icon}
           <div>
-            <S.Title>{project.title}</S.Title>
-            <S.Description>
+            <h2 className={styles.title}>{project.title}</h2>
+            <p className={styles.description}>
               {t(`projects.${project.title.replace(/ /g, '-').toLowerCase()}`)}
-            </S.Description>
+            </p>
           </div>
         </span>
-      </S.Container>
-    </S.Wrapper>
+      </div>
+    </motion.a>
   )
 }
