+++ src/domains/projects/TopProjects/styles.module.css
@@ -1,0 +1,10 @@
+.grid {
+  display: grid;
+  grid-template-columns: repeat(auto-fit, minmax(200px, 1fr));
+  grid-auto-flow: row;
+  padding-top: var(--space-4);
+}
+
+.grid > div > div {
+  height: auto;
+}
