--- src/domains/projects/project-item/index.ts
@@ -1,1 +1,0 @@
-export { ProjectItem } from './project.component'
