--- src/domains/projects/styles.ts
@@ -1,10 +1,0 @@
-import { styled } from 'src/ui/theme'
-
-export const Wrapper = styled('section', {
-  display: 'flex',
-  flexDirection: 'column',
-  justifyContent: 'center',
-  alignItems: 'flex-start',
-  transition: '$transitonTheme',
-  padding: '2rem 0',
-})
